+++ all.sh
@@ -1,4 +1,12 @@
 #!/bin/bash
+#
+# USAGE
+# -----
+#
+#   ./all.sh pub get
+#
+#   ./all.sh pub run test
+#
 
 set -e
 cd `dirname $0`
@@ -7,15 +15,20 @@ ARGS="${@:1}"
 
 visit() {
   NAME=$1
+  echo ""
   echo "-------------------------------------------------"
-  echo "Getting dependencies for '$NAME'"
+  echo "'$NAME': $ARGS"
   echo "-------------------------------------------------"
+  echo ""
   cd $NAME
-  pub get
+
+  # Run the command
+  $ARGS
+
   cd ..
+  echo ""
 }
 
 visit cryptography
 visit kms
-
 visit kms_adapter_cupertino