+++ cryptography/lib/src/cryptography/private_key.dart
@@ -26,43 +26,75 @@ import 'package:cryptography/utils.dart';
 /// final keyPair = x25519.keyPairGenerator.generateSync();
 /// final privateKey = keyPair.privateKey;
 /// ```
-class PrivateKey {
+abstract class PrivateKey {
   static final _random = Random.secure();
 
-  /// Bytes of the key. May be null.
-  final List<int> bytes;
+  /// Constructs a private key with the bytes.
+  const factory PrivateKey(List<int> bytes) = _PrivateKey;
 
-  const PrivateKey(this.bytes) : assert(bytes != null);
+  /// Constructor for subclasses.
+  const PrivateKey.constructor();
 
   /// Generates N random bytes.
   ///
   /// You can optionally give a random number generator that's used.
   factory PrivateKey.randomBytes(int length, {Random random}) {
-    random ??= PrivateKey._random;
+    random ??= _random;
     final data = Uint8List(length);
-    for (var i = 0; i < data.length; i++) {
+    for (var i = 0; i < length; i++) {
       data[i] = random.nextInt(256);
     }
     return PrivateKey(data);
   }
 
+  @Deprecated('Use `extractSync()` instead')
+  List<int> get bytes => extractSync();
+
+  /// Extracts the bytes.
+  ///
+  /// Throws [UnsupportedError] if extraction is not supported.
+  Future<List<int>> extract() => Future<List<int>>(() => extractSync());
+
+  /// Extracts the bytes synchronously.
+  /// Throws [UnsupportedError] if extraction is not supported.
+  List<int> extractSync();
+}
+
+class _PrivateKey extends PrivateKey {
+  /// Bytes of the key. May be null.
+  final List<int> _bytes;
+
+  const _PrivateKey(this._bytes)
+      : assert(_bytes != null),
+        super.constructor();
+
   @override
   int get hashCode {
-    /// Exposes maximum 16 bits of the key.
     var h = 0;
-    final bytes = this.bytes;
+    final bytes = _bytes;
     for (var i = 0; i < bytes.length; i++) {
       final b = bytes[i];
-      h ^= (b << (i % 16)) ^ (b >> (16 - (i % 16)));
+
+      // Exposes at most 31 bits
+      h = 0x7FFFFFFF & ((31 * h) ^ b);
     }
+
+    // For short values, expose max 15 bits.
+    if (bytes.length < 8) {
+      h = 0x7FFF & h;
+    }
+
     return h;
   }
 
   @override
   bool operator ==(other) {
-    return other is PrivateKey &&
-        constantTimeBytesEquality.equals(bytes, other.bytes);
+    return other is _PrivateKey &&
+        constantTimeBytesEquality.equals(_bytes, other._bytes);
   }
+
+  @override
+  List<int> extractSync() => _bytes;
 
   @override
   String toString() => 'PrivateKey(...)';