+++ cryptography/test/src/cryptography/algorithms/ec_nist_test.dart
@@ -13,7 +13,6 @@
 // limitations under the License.
 
 import 'package:cryptography/cryptography.dart';
-import 'package:cryptography/src/utils/hex.dart';
 import 'package:test/test.dart';
 
 void main() {