+++ cryptography/test/src/cryptography/algorithms/poly1305_test.dart
@@ -77,8 +77,8 @@ void main() {
         nonce: nonce,
       );
       expect(
-        hexFromBytes(poly1305Key.bytes),
-        hexFromBytes(expectedPoly1305Key.bytes),
+        hexFromBytes(poly1305Key.extractSync()),
+        hexFromBytes(expectedPoly1305Key.extractSync()),
       );
     });
   });