+++ cryptography/test/src/cryptography/algorithms/sha3_test.dart
@@ -12,7 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import 'package:crypto/crypto.dart' as google_crypto;
 import 'package:cryptography/cryptography.dart';
 import 'package:cryptography/utils.dart';
 import 'package:test/test.dart';