+++ test.sh
@@ -3,18 +3,7 @@
 set -e
 cd `dirname $0`
 
-ARGS="${@:1}"
-
-visit() {
-  NAME=$1
-  echo "-------------------------------------------------"
-  echo "Testing '$NAME'"
-  echo "-------------------------------------------------"
-  cd $NAME
-  pub run test $ARGS
-  cd ..
-}
-
-visit cryptography
-visit kms
-visit kms_adapter_cupertino
+./all.sh pub get
+./all.sh pub run test
+./all.sh flutter pub get
+./all.sh flutter pub run test --platform=vm